C04

+++ C04/ex03/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: sgang <xifoxy.ru@gmail.com>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/09 15:50:09 by sgang             #+#    #+#             */
-/*   Updated: 2020/07/09 16:06:29 by sgang            ###   ########.fr       */
+/*   Updated: 2020/07/14 17:41:38 by sgang            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ int	is_white_space(char ch)
 
 int	ft_atoi(char *str)
 {
-	int ret;
-	int mark;
+	long long	ret;
+	int			mark;
 
 	ret = 0;
 	mark = 1;
